fix pACNC
There used to be a possible infinite loop when looking for no clobber
due to the map searching for `free` metadatas instead of "available"
metadatas. Because of this search the final sentinel group needs to not
include tombstones anymore and instead is filled with the "used"
metadata slice of splatted `0b10101010`.